+++ pom.xml
@@ -3,9 +3,9 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>es.ua.mads</groupId>
-    <artifactId>mads-todolist</artifactId>
-    <version>1.0.0</version>
+    <groupId>adel.es</groupId>
+    <artifactId>AB-ToDoLiSt</artifactId>
+    <version>1.0.1-SNAPSHOT</version>
 
     <parent>
         <groupId>org.springframework.boot</groupId>
