--- src/main/java/madstodolist/controller/exception/UsuarioNoLogeadoException.java
@@ -1,8 +1,0 @@
-package madstodolist.controller.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(value = HttpStatus.UNAUTHORIZED, reason="Usuario no autorizado")
-public class UsuarioNoLogeadoException extends RuntimeException {
-}
