--- src/main/java/madstodolist/repository/TareaRepository.java
@@ -1,7 +1,0 @@
-package madstodolist.repository;
-
-import madstodolist.model.Tarea;
-import org.springframework.data.repository.CrudRepository;
-
-public interface TareaRepository extends CrudRepository<Tarea, Long> {
-}
