--- src/main/java/madstodolist/repository/UsuarioRepository.java
@@ -1,10 +1,0 @@
-package madstodolist.repository;
-
-import madstodolist.model.Usuario;
-import org.springframework.data.repository.CrudRepository;
-
-import java.util.Optional;
-
-public interface UsuarioRepository extends CrudRepository<Usuario, Long> {
-    Optional<Usuario> findByEmail(String s);
-}
