--- src/main/java/madstodolist/service/TareaServiceException.java
@@ -1,8 +1,0 @@
-package madstodolist.service;
-
-public class TareaServiceException extends RuntimeException {
-
-    public TareaServiceException(String message) {
-        super(message);
-    }
-}
