--- src/main/java/madstodolist/service/UsuarioServiceException.java
@@ -1,8 +1,0 @@
-package madstodolist.service;
-
-public class UsuarioServiceException extends RuntimeException {
-
-    public UsuarioServiceException(String message) {
-        super(message);
-    }
-}
