+++ src/main/java/todolist/controller/exception/TareaNotFoundException.java
@@ -1,0 +1,8 @@
+package todolist.controller.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(value = HttpStatus.NOT_FOUND, reason="Tarea no encontrada")
+public class TareaNotFoundException extends RuntimeException {
+}
