+++ src/main/java/todolist/controller/exception/UsuarioNoLogeadoException.java
@@ -1,0 +1,8 @@
+package todolist.controller.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(value = HttpStatus.UNAUTHORIZED, reason="Usuario no autorizado")
+public class UsuarioNoLogeadoException extends RuntimeException {
+}
