+++ src/main/java/todolist/dto/LoginData.java
@@ -1,0 +1,23 @@
+package todolist.dto;
+
+// Clase de datos para el formulario de login
+public class LoginData {
+    private String eMail;
+    private String password;
+
+    public String geteMail() {
+        return eMail;
+    }
+
+    public void seteMail(String eMail) {
+        this.eMail = eMail;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+}
