+++ src/main/java/todolist/repository/TareaRepository.java
@@ -1,0 +1,7 @@
+package todolist.repository;
+
+import todolist.model.Tarea;
+import org.springframework.data.repository.CrudRepository;
+
+public interface TareaRepository extends CrudRepository<Tarea, Long> {
+}
