+++ src/main/java/todolist/repository/UsuarioRepository.java
@@ -1,0 +1,10 @@
+package todolist.repository;
+
+import todolist.model.Usuario;
+import org.springframework.data.repository.CrudRepository;
+
+import java.util.Optional;
+
+public interface UsuarioRepository extends CrudRepository<Usuario, Long> {
+    Optional<Usuario> findByEmail(String s);
+}
