+++ src/main/java/todolist/service/TareaServiceException.java
@@ -1,0 +1,8 @@
+package todolist.service;
+
+public class TareaServiceException extends RuntimeException {
+
+    public TareaServiceException(String message) {
+        super(message);
+    }
+}
