+++ src/main/java/todolist/service/UsuarioServiceException.java
@@ -1,0 +1,8 @@
+package todolist.service;
+
+public class UsuarioServiceException extends RuntimeException {
+
+    public UsuarioServiceException(String message) {
+        super(message);
+    }
+}
