+++ src/main/resources/application.properties
@@ -3,7 +3,7 @@ spring.datasource.url=jdbc:h2:mem:dev
 spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.H2Dialect
 spring.jpa.hibernate.ddl-auto=update
 logging.level.org.hibernate.SQL=debug
-logging.level.madstodolist=debug
+logging.level.todolist=debug
 spring.sql.init.mode=never
 spring.h2.console.enabled=true
 spring.h2.console.path=/h2-console
